Fixed code over comments
