+++ bring/src/main/java/com/bobocode/petros/bring/scanner/impl/DefaultBeanPostProcessorScanner.java
@@ -5,8 +5,8 @@ import com.bobocode.petros.bring.scanner.BeanPostProcessorScanner;
 import lombok.SneakyThrows;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
-import java.util.stream.Stream;
 
 /**
  * Implementation of {@link BeanPostProcessorScanner}.
@@ -23,7 +23,7 @@ public class DefaultBeanPostProcessorScanner implements BeanPostProcessorScanner
     @Override
     public List<BeanPostProcessor> scan(Set<Class<?>> classes) {
         return classes.stream()
-                .flatMap(Stream::ofNullable)
+                .filter(Objects::nonNull)
                 .filter(BeanPostProcessor.class::isAssignableFrom)
                 .map(this::createInstance)
                 .toList();
