--- bring/src/test/java/com/bobocode/petros/bring/scanner/mocks/GreetingService.java
@@ -1,5 +1,0 @@
-package com.bobocode.petros.bring.scanner.mocks;
-
-public interface GreetingService {
-    void greeting();
-}
