+++ bring/src/test/java/com/bobocode/petros/bring/scanner/mocks/postprocessor/EveningBeanPostProcessor.java
@@ -1,16 +1,21 @@
-package com.bobocode.petros.bring.scanner.mocks;
+package com.bobocode.petros.bring.scanner.mocks.postprocessor;
 
 import com.bobocode.petros.bring.context.domain.BeanReference;
 import com.bobocode.petros.bring.factory.postprocessor.BeanPostProcessor;
 
-public class MorningBeanPostProcessor implements BeanPostProcessor {
+public class EveningBeanPostProcessor implements BeanPostProcessor {
     @Override
     public BeanReference postProcessBeforeInitialization(BeanReference beanReference) {
-        return null;
+        return beanReference;
     }
 
     @Override
     public BeanReference postProcessAfterInitialization(BeanReference beanReference) {
-        return null;
+        return beanReference;
+    }
+
+    @Override
+    public int getOrder() {
+        return 0;
     }
 }
