+++ bring/src/test/java/com/bobocode/petros/bring/scanner/mocks/postprocessor/EveningService.java
@@ -1,4 +1,4 @@
-package com.bobocode.petros.bring.scanner.mocks;
+package com.bobocode.petros.bring.scanner.mocks.postprocessor;
 
 public class EveningService implements GreetingService{
     @Override
