+++ bring/src/test/java/com/bobocode/petros/bring/scanner/mocks/postprocessor/GreetingService.java
@@ -1,0 +1,5 @@
+package com.bobocode.petros.bring.scanner.mocks.postprocessor;
+
+public interface GreetingService {
+    void greeting();
+}
