+++ bring/src/test/java/com/bobocode/petros/bring/scanner/mocks/postprocessor/MorningBeanPostProcessor.java
@@ -1,21 +1,16 @@
-package com.bobocode.petros.bring.scanner.mocks;
+package com.bobocode.petros.bring.scanner.mocks.postprocessor;
 
 import com.bobocode.petros.bring.context.domain.BeanReference;
 import com.bobocode.petros.bring.factory.postprocessor.BeanPostProcessor;
 
-public class EveningBeanPostProcessor implements BeanPostProcessor {
+public class MorningBeanPostProcessor implements BeanPostProcessor {
     @Override
     public BeanReference postProcessBeforeInitialization(BeanReference beanReference) {
-        return beanReference;
+        return null;
     }
 
     @Override
     public BeanReference postProcessAfterInitialization(BeanReference beanReference) {
-        return beanReference;
-    }
-
-    @Override
-    public int getOrder() {
-        return 0;
+        return null;
     }
 }
